TD-Jour6/products-api/requirements.txt
fastapi==0.115.2
uvicorn[standard]==0.32.0
pydantic==2.9.2
pyjwt==2.11.0
python-multipart==0.0.9
